Give login flows their own two modules.
The interfaces and the parameter helper were sitting in base.ts only
because a flow cannot import from the module that lists the flows. That
is a reason to split the list off, not a reason to put login flows
among the core service abstractions.

loginFlow.ts now says what a flow is; loginFlowRegistry.ts says which
ones exist. Flows import the first, the registry imports both, and
nothing imports the registry except the CLI and the config loader — so
the cycle cannot come back.
