+++ src/services/core/base.ts
@@ -3,7 +3,7 @@
  */
 
 import type { Browser, BrowserContext, Page, Response } from 'playwright';
-import type { z, ZodError, ZodType, ZodTypeAny } from 'zod';
+import type { z, ZodError, ZodTypeAny } from 'zod';
 import {
   ApiCredentialStatus,
   ApiCredentials,
@@ -547,65 +547,4 @@ export abstract class BrowserFollowupServiceSession extends ServiceSession {
     );
     return this.performBrowserFollowup(context, oldCredentials);
   }
-}
-
-/**
- * Thrown when the parameters stored for a login flow do not match its schema.
- */
-export class LoginFlowParamsInvalidError extends Error {
-  constructor(flowName: string, detail: string) {
-    super(`Invalid parameters for login flow '${flowName}': ${detail}`);
-    this.name = 'LoginFlowParamsInvalidError';
-  }
-}
-
-/**
- * A generic browser login that a user-registered service can opt into,
- * configured with the parameters that service was registered with.
- *
- * A built-in service implements its login in code. A registered service picks a
- * flow by name and supplies parameters as JSON, so new flows become available to
- * every registered service without the CLI knowing anything about them.
- */
-export interface LoginFlow {
-  /** Sentence describing this configuration, for `services info`. */
-  describe(loginUrl: string): string;
-  createSession(service: Service, appNamePrefix: string): ServiceSession;
-}
-
-/**
- * The class side of a flow: the kind of automation, before anyone configures
- * it. Its instances are the configured flows.
- *
- * The metadata has to be static, since flows are listed and described before
- * any parameters exist — and TypeScript has no `abstract static`, so this
- * interface is what holds a flow class to declaring it. A flow states the
- * obligation with `satisfies LoginFlowClass`; the registry checks it again by
- * being typed as an array of these.
- */
-export interface LoginFlowClass {
-  new (params: unknown): LoginFlow;
-  /** Value of `--login-flow`. Not `name`, which every class already has. */
-  readonly flowName: string;
-  /** One-line explanation, listed by the CLI when the flow name is wrong. */
-  readonly summary: string;
-  /** Reference documentation, shown by `services register --help`. */
-  readonly details: string;
-  readonly paramsSchema: ZodTypeAny;
-}
-
-/**
- * Validate the parameters a flow was registered with, against the flow's own
- * schema. Inference gives back the schema's type, so a flow storing the result
- * cannot disagree with the schema it declared.
- */
-export function parseLoginFlowParams<Params>(
-  flowClass: { readonly flowName: string; readonly paramsSchema: ZodType<Params> },
-  rawParams: unknown
-): Params {
-  const result = flowClass.paramsSchema.safeParse(rawParams ?? {});
-  if (!result.success) {
-    throw new LoginFlowParamsInvalidError(flowClass.flowName, describeSchemaIssues(result.error));
-  }
-  return result.data;
 }
