+++ src/services/core/registered.ts
@@ -8,7 +8,8 @@
  */
 
 import { ApiCredentialStatus, type ApiCredentials } from '../../apiCredentials/base.js';
-import { Service, type LoginFlow, type ServiceSession } from './base.js';
+import { Service, type ServiceSession } from './base.js';
+import type { LoginFlow } from './loginFlow.js';
 
 /**
  * Where a registered service gets its browser login, if anywhere. The two
